moved queue executer into command_dispatcher
