--- .vscode/launch.json
@@ -1,22 +1,0 @@
-{
-        "version": "0.2.0",
-        "configurations": [
-        {
-            "type": "node",
-            "request": "launch",
-            "name": "Launch via NPM",
-            "runtimeExecutable": "npm",
-            "runtimeArgs": [
-                "run-script",
-                "debug"
-            ],
-            "port": 9229
-        },
-            {
-                "type": "node",
-                "request": "launch",
-                "name": "Launch Program",
-                "program": "${workspaceRoot}\\index.js"
-            }
-        ]
-    }
